+++ tensorflow_datasets/core/download/extractor_test.py
@@ -22,7 +22,7 @@ from __future__ import print_function
 import os
 
 from absl.testing import absltest
-import tensorflow as tf
+import tensorflow.compat.v2 as tf
 from tensorflow_datasets import testing
 from tensorflow_datasets.core.download import extractor
 from tensorflow_datasets.core.download import resource as resource_lib