+++ tensorflow_datasets/core/features/feature.py
@@ -92,7 +92,7 @@ import collections
 
 import numpy as np
 import six
-import tensorflow as tf
+import tensorflow.compat.v2 as tf
 
 from tensorflow_datasets.core import api_utils
 from tensorflow_datasets.core import utils