+++ tensorflow_datasets/core/features/text/subword_text_encoder_test.py
@@ -24,7 +24,7 @@ from __future__ import unicode_literals
 import os
 
 from absl.testing import parameterized
-import tensorflow as tf
+import tensorflow.compat.v2 as tf
 from tensorflow_datasets import testing
 from tensorflow_datasets.core.features.text import subword_text_encoder
 from tensorflow_datasets.core.features.text import text_encoder