+++ tensorflow_datasets/testing/oxford_iiit_pet.py
@@ -27,7 +27,7 @@ import tempfile
 from absl import app
 from absl import flags
 
-import tensorflow as tf
+import tensorflow.compat.v2 as tf
 
 from tensorflow_datasets.core.utils import py_utils
 from tensorflow_datasets.testing import fake_data_utils