+++ docs/release_notes.md
@@ -18,3 +18,5 @@
 *   Add e-SNLI dataset from the paper
     [e-SNLI](http://papers.nips.cc/paper/8163-e-snli-natural-language-inference-with-natural-language-explanations.pdf).
 *   Add [Opinosis dataset](https://www.aclweb.org/anthology/C10-1039.pdf).
+*   Add SCAN dataset introduced
+    [here](https://arxiv.org/pdf/1711.00350.pdf).